pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnimarket"
version = "0.1.0"
description = "OmniMarket - Portable ONEX workflow package registry"
authors = [
    {name = "OmniNode.ai", email = "contact@omninode.ai"}
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
keywords = [
    "onex",
    "marketplace",
    "workflow",
    "portable",
    "contract",
    "event-driven",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "omnibase_core>=0.39.0",
    "pydantic>=2.12.5,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "aiokafka>=0.12.0",
    "asyncpg>=0.30.0",
    "python-dateutil>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/OmniNode-ai/omnimarket"
Repository = "https://github.com/OmniNode-ai/omnimarket"

[project.entry-points."onex.nodes"]
node_aislop_sweep = "omnimarket.nodes.node_aislop_sweep"
node_merge_sweep = "omnimarket.nodes.node_merge_sweep"
node_platform_readiness = "omnimarket.nodes.node_platform_readiness"
node_build_loop = "omnimarket.nodes.node_build_loop"
node_ticket_pipeline = "omnimarket.nodes.node_ticket_pipeline"
node_close_out = "omnimarket.nodes.node_close_out"
node_hostile_reviewer = "omnimarket.nodes.node_hostile_reviewer"
node_local_review = "omnimarket.nodes.node_local_review"
node_pr_polish = "omnimarket.nodes.node_pr_polish"
node_dod_verify = "omnimarket.nodes.node_dod_verify"
node_coverage_sweep = "omnimarket.nodes.node_coverage_sweep"
node_compliance_sweep = "omnimarket.nodes.node_compliance_sweep"
node_dashboard_sweep = "omnimarket.nodes.node_dashboard_sweep"
node_data_flow_sweep = "omnimarket.nodes.node_data_flow_sweep"
node_golden_chain_sweep = "omnimarket.nodes.node_golden_chain_sweep"
node_runtime_sweep = "omnimarket.nodes.node_runtime_sweep"
node_design_to_plan = "omnimarket.nodes.node_design_to_plan"
node_ticket_work = "omnimarket.nodes.node_ticket_work"
node_create_ticket = "omnimarket.nodes.node_create_ticket"
node_plan_to_tickets = "omnimarket.nodes.node_plan_to_tickets"
node_release = "omnimarket.nodes.node_release"
node_redeploy = "omnimarket.nodes.node_redeploy"
node_log_projection = "omnimarket.nodes.node_log_projection"
node_data_verification = "omnimarket.nodes.node_data_verification"
node_process_watchdog = "omnimarket.nodes.node_process_watchdog"

[tool.uv.sources]
omnibase-core = { path = "../omnibase_core", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/omnimarket"]

[tool.hatch.build]
artifacts = ["src/omnimarket/**/*.yaml"]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13,<0.16.0",
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250822",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF", "C4", "T20", "PT", "ARG", "SLF", "RET"]
ignore = [
    "E501",      # Line too long (will fix incrementally)
    "ARG001",    # Unused function argument
    "ARG002",    # Unused method argument
    "RET504",    # Unnecessary assign before return
    "RUF012",    # Mutable class default (ClassVar annotation)
    "SIM103",    # Needless bool return
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["T201", "SLF001"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
python_version = "3.12"
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = false
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests that test individual components in isolation",
    "integration: Integration tests that test multiple components together",
    "slow: Tests that take a long time to run (>1s)",
]
